requirements.txt
torch
transformers
pytesseract
opencv-python
Pillow
git+https://github.com/facebookresearch/segment-anything.git